examples/requirements.txt
git+https://github.com/knowsuchagency/orkestra.git@main#egg=orkestra[powertools]
# orkestra[powertools]==0.3.3
fastapi==0.65.1
mangum==0.11.0
boto3==1.17.18
python-dotenv==0.17.1
